src/main/python/requirements.txt
flask
mysql-connector-python
pika
